requirements.txt
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.6
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.8
nltk==3.4.5
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.4
python-rake==1.4.5